+++ public/js/controllers/ActivitiesCtrl.js
@@ -26,31 +26,46 @@ angular
                 filterValue = null;
             }
 
+            $scope.keyCounter = function (objIn) {
+                return Object.keys(objIn).length;
+            };
+
+            $scope.getGroupItems = function (values) {
+                var returnArray = [];
+                Object.keys(values).forEach(function (key) {
+                    returnArray.push(values[key]);
+                });
+
+                return returnArray;
+            };
+
             sActivity
                 .getActivities($scope.activitiesOffset, $scope.activitiesLimit, filterValue)
                 .then(function (activities) {
                     $scope.app.unreadActivitiesCount = 0;
-                    activities.forEach(function (activity, key) {
-                        if (activity.data.type === 'View' && activity.data.object && activity.data.object['@type'] === 'Activity') {
-                            if (!lastViewTime || activity.updatedAt > lastViewTime) {
-                                lastViewTime = activity.updatedAt;
+                    activities.forEach(function (activityGroups, groupKey) {
+                        Object.keys(activityGroups.values).forEach(function (key) {
+                            var activity = activityGroups.values[key];
+                            if (activity.data.type === 'View' && activity.data.object && activity.data.object['@type'] === 'Activity') {
+                                if (!lastViewTime || activity.updatedAt > lastViewTime) {
+                                    lastViewTime = activity.updatedAt;
+                                }
+                                activities.splice(groupKey, 1);
+                            } else if (!lastViewTime || activity.updatedAt > lastViewTime) {
+                                activity.isNew = '-new';
                             }
-                            activities.splice(key, 1);
-                        } else if (!lastViewTime || activity.updatedAt > lastViewTime) {
-                            activity.isNew = '-new';
-                        }
+                        });
                     });
-
-                    $scope.showLoadMoreActivities = !(activities.length < $scope.activitiesLimit);
+                    $scope.showLoadMoreActivities = (!activities.length < $scope.activitiesLimit);
                     $scope.activities = $scope.activities.concat(activities);
 
                     var element = angular.element($document[0].getElementsByClassName('lightbox_content'));
                     if (element && element[0] && element[0].scrollHeight) {
-                        $scope.$watch(element[0].scrollHeight, function () {
+                       /* $scope.$watch(element[0].scrollHeight, function () {
                             if (activities.length && element[0].clientHeight >= element[0].scrollHeight) {
-                                $scope.loadActivities();
+                               // $scope.loadActivities();
                             }
-                        }, true);
+                        }, true);*/
                     }
 
                 });